gcp-calorie/requirements.txt
# Function dependencies, for example:
# package>=version
boto3
google-cloud-storage
pandas
fsspec
gcsfs
google-cloud-pubsub
openpyxl
beautifulsoup4
requests